Split lib file in utils folder, fix import
