+++ src/app/page.tsx
@@ -1,7 +1,5 @@
 import { NextPage } from 'next';
 
-export const runtime = 'experimental-edge';
-
 const PageIndex: NextPage = () => {
   return null;
 };
