+++ src/app/tx/[tx]/loading.tsx
@@ -1,0 +1,8 @@
+// Displayed when /tx/[tx] is loading
+export default function Loading() {
+  return (
+    <div className="resultContainer">
+      <div className="receipt-text">Loading...</div>
+    </div>
+  );
+}
